+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/model/builder/CategoryRarityModelBuilder.java
@@ -13,15 +13,15 @@ public class CategoryRarityModelBuilder implements IModelBuilder {
 
     @Override
     public Model build(Object modelBuilderData) {
-        Map<String, Integer> featureValueToCountMap = castModelBuilderData(modelBuilderData);
+        Map<String, Long> featureValueToCountMap = castModelBuilderData(modelBuilderData);
         return new CategoryRarityModel(getOccurrencesToNumOfFeatures(featureValueToCountMap));
     }
 
 
-    protected Map<Integer, Double> getOccurrencesToNumOfFeatures(Map<String, Integer> featureValueToCountMap) {
-        Map<Integer, Double> occurrencesToNumOfFeatures = new HashMap<>();
-        for (Map.Entry<String, Integer> entry : featureValueToCountMap.entrySet()) {
-            int count = entry.getValue();
+    protected Map<Long, Double> getOccurrencesToNumOfFeatures(Map<String, Long> featureValueToCountMap) {
+        Map<Long, Double> occurrencesToNumOfFeatures = new HashMap<>();
+        for (Map.Entry<String, Long> entry : featureValueToCountMap.entrySet()) {
+            long count = entry.getValue();
             Double numOfFeatures = occurrencesToNumOfFeatures.get(count);
             if (numOfFeatures == null) {
                 numOfFeatures = 0D;
@@ -33,7 +33,7 @@ public class CategoryRarityModelBuilder implements IModelBuilder {
 
 
     @SuppressWarnings("unchecked")
-    protected Map<String, Integer> castModelBuilderData(Object modelBuilderData) {
+    protected Map<String, Long> castModelBuilderData(Object modelBuilderData) {
         if (modelBuilderData == null) {
             throw new IllegalArgumentException();
         }
@@ -42,6 +42,6 @@ public class CategoryRarityModelBuilder implements IModelBuilder {
             logger.error(errorMsg);
             throw new IllegalArgumentException(errorMsg);
         }
-        return (Map<String, Integer>) modelBuilderData;
+        return (Map<String, Long>)modelBuilderData;
     }
 }
