+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/model/builder/CategoryRarityModelWithFeatureOccurrencesDataBuilder.java
@@ -7,7 +7,7 @@ import java.util.Map;
 public class CategoryRarityModelWithFeatureOccurrencesDataBuilder extends CategoryRarityModelBuilder {
     @Override
     public Model build(Object modelBuilderData) {
-        Map<String, Integer> featureValueToCountMap = castModelBuilderData(modelBuilderData);
+        Map<String, Long> featureValueToCountMap = castModelBuilderData(modelBuilderData);
         return new CategoryRarityModelWithFeatureOccurrencesData(getOccurrencesToNumOfFeatures(featureValueToCountMap));
     }
 }
